+++ ikuzo/service/x/imageproxy/service.go
@@ -20,61 +20,36 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"log"
 	"net/http"
-	"path/filepath"
+	"net/url"
+	"os/exec"
 	"strings"
 	"time"
 
-	"github.com/go-chi/chi"
-	"github.com/rs/zerolog/log"
+	"github.com/rs/zerolog"
+	"golang.org/x/sync/singleflight"
+
+	lru "github.com/hashicorp/golang-lru"
 )
 
-type Option func(*Service) error
+// var _ domain.Service = (*Service)(nil)
 
 type Service struct {
-	client      http.Client
-	cacheDir    string // The path to the imageCache
-	timeOut     int    // timelimit for request served by this proxy. 0 is for no timeout
-	proxyPrefix string // The prefix where we mount the imageproxy. default: imageproxy. default: imageproxy.
-	memoryCache string
-	referrers   []string
-	blacklist   []string
-	// deepzoom    bool     // Enable deepzoom of remote images.
-}
-
-func SetCacheDir(path string) Option {
-	return func(s *Service) error {
-		s.cacheDir = path
-		return nil
-	}
-}
-
-func SetTimeout(duration int) Option {
-	return func(s *Service) error {
-		s.timeOut = duration
-		return nil
-	}
-}
-
-func SetProxyReferrer(referrer []string) Option {
-	return func(s *Service) error {
-		s.referrers = referrer
-		return nil
-	}
-}
-
-func SetBlackList(blacklist []string) Option {
-	return func(s *Service) error {
-		s.blacklist = blacklist
-		return nil
-	}
-}
-
-func SetProxyPrefix(prefix string) Option {
-	return func(s *Service) error {
-		s.proxyPrefix = prefix
-		return nil
-	}
+	client          http.Client
+	lruCache        *lru.ARCCache
+	cacheDir        string // The path to the imageCache
+	maxSizeCacheDir int    //  max size of the cache directory on disK
+	timeOut         int    // timelimit for request served by this proxy. 0 is for no timeout
+	proxyPrefix     string // The prefix where we mount the imageproxy. default: imageproxy. default: imageproxy.
+	referrers       []string
+	allowList       []string
+	refuselist      []string
+	m               Metrics
+	log             zerolog.Logger
+	enableResize    bool
+	singleSetCache  singleflight.Group
+	// orgs         domain.OrgConfigRetriever
 }
 
 func NewService(options ...Option) (*Service, error) {
@@ -82,7 +57,7 @@ func NewService(options ...Option) (*Service, error) {
 		cacheDir:    "/tmp/imageproxy",
 		timeOut:     10,
 		proxyPrefix: "imageproxy",
-		memoryCache: "memory:500:1h",
+		log:         zerolog.Nop(),
 	}
 
 	// apply options
@@ -92,52 +67,200 @@ func NewService(options ...Option) (*Service, error) {
 		}
 	}
 
+	if s.enableResize {
+		s.enableResize = s.checkForVips()
+	}
+
 	s.client = http.Client{Timeout: time.Duration(s.timeOut) * time.Second}
 
 	return s, nil
 }
 
-func (s *Service) Routes() chi.Router {
-	router := chi.NewRouter()
+func (s *Service) checkForVips() bool {
+	_, err := exec.LookPath("vips")
+	if err != nil {
+		s.log.Warn().Msg("libvips is not installed so disabling imageproxy resize options")
+		return false
+	}
 
-	proxyPrefix := fmt.Sprintf("/%s/{options}/*", s.proxyPrefix)
-	router.Get(proxyPrefix, s.proxyImage)
+	return true
+}
 
-	return router
+func deepZoomExternally(from string) error {
+	cleanFrom := strings.TrimSuffix(from, ".dzi")
+	args := []string{
+		"dzsave",
+		cleanFrom,
+		cleanFrom,
+	}
+
+	path, err := exec.LookPath("vips")
+	if err != nil {
+		return err
+	}
+
+	cmd := exec.Command(path, args...)
+
+	log.Printf("deepzoom command: %s", cmd.String())
+
+	return cmd.Run()
+}
+
+func resizeExternally(from, to, size string) error {
+	args := []string{
+		"--size", size,
+		"--output", to,
+		from,
+	}
+
+	path, err := exec.LookPath("vipsthumbnail")
+	if err != nil {
+		return err
+	}
+
+	cmd := exec.Command(path, args...)
+
+	return cmd.Run()
 }
 
 func (s *Service) Do(ctx context.Context, req *Request, w io.Writer) error {
 	_ = ctx
-	cachePath := filepath.Join(s.cacheDir, req.sourcePath())
+
+	if s.lruCache != nil {
+		if s.lruCache.Contains(req.CacheKey) {
+			data, ok := s.lruCache.Get(req.CacheKey)
+			if ok {
+				req.CacheType = Lru
+
+				_, err := fmt.Fprintf(w, "%s", data)
+
+				s.m.IncLruCache()
+
+				return err
+			}
+		}
+	}
+
+	isCached := existsInCache(req.cacheKeyPath())
+
+	if !isCached {
+		hasSource := existsInCache(req.downloadedSourcePath())
+		if !hasSource {
+			_, err, shared := s.singleSetCache.Do(
+				req.SourceURL,
+				func() (interface{}, error) {
+					s.log.Info().Str("path", req.downloadedSourcePath()).Msg("started storing source")
+					err := s.storeSource(req)
+					if err != nil {
+						s.log.Error().Err(err).Str("url", req.SourceURL).
+							Str("sourcePath", req.downloadedSourcePath()).
+							Str("storePath", req.cacheKeyPath()).
+							Msgf("unexpected error saving source; %s", err)
+						return nil, err
+					}
+					s.log.Info().Str("path", req.downloadedSourcePath()).Msg("finished storing source")
+					return nil, nil
+				},
+			)
+
+			if err != nil {
+				return err
+			}
+
+			if !shared {
+				req.CacheType = Source
+			}
+
+			hasSource = true
+		}
+
+		if !isCached && req.thumbnailOpts != "" && hasSource {
+			_, err, _ := s.singleSetCache.Do(
+				req.CacheKey,
+				func() (interface{}, error) {
+					return nil, resizeExternally(req.downloadedSourcePath(), req.cacheKeyPath(), req.thumbnailOpts)
+				},
+			)
+			if err != nil {
+				s.log.Error().Err(err).Str("url", req.SourceURL).
+					Str("sourcePath", req.downloadedSourcePath()).
+					Str("storePath", req.cacheKeyPath()).
+					Msgf("unexpected error creating thumbnail; %s", err)
+
+				req.CacheKey = encodeURL(req.SourceURL)
+			}
+
+			s.m.IncResize()
+		}
+
+		if !isCached && req.SubPath == deepZoomSuffix {
+			_, err, _ := s.singleSetCache.Do(
+				req.CacheKey,
+				func() (interface{}, error) {
+					return nil, deepZoomExternally(req.downloadedSourcePath())
+				},
+			)
+			if err != nil {
+				s.log.Error().Err(err).Str("url", req.SourceURL).Msg("unexpected error creating deepzoom")
+				s.m.IncError()
+
+				return err
+			}
+
+			s.m.IncDeepZoom()
+		}
+	}
+
 	// check cache
-	r, err := req.Read(cachePath)
+	r, err := req.Read(req.cacheKeyPath())
 	if err != nil && !errors.Is(err, ErrCacheKeyNotFound) {
-		log.Error().Err(err).Str("cmp", "imageproxy").Msg("unexpected error reading from cache")
+		s.log.Error().Err(err).Str("url", req.SourceURL).Msg("unexpected error reading from cache")
+		s.m.IncError()
+
 		return err
 	}
 
 	if !errors.Is(err, ErrCacheKeyNotFound) {
 		defer r.Close()
 
-		_, err = io.Copy(w, r)
+		var buf bytes.Buffer
+		tee := io.TeeReader(r, &buf)
+
+		_, err = io.Copy(w, tee)
 		if err != nil {
-			log.Error().Err(err).Str("cmp", "imageproxy").Msg("error copying image from cache")
+			s.log.Error().Err(err).Str("url", req.SourceURL).Msg("error copying data from cache")
+			s.m.IncError()
+
 			return err
 		}
 
-		return nil
+		if req.CacheType == "" {
+			req.CacheType = Cache
+		}
+
+		s.lruCache.Add(req.CacheKey, buf.Bytes())
+
+		s.m.IncCache()
 	}
 
+	return nil
+}
+
+func (s *Service) storeSource(req *Request) error {
 	// make request
 	proxyRequest, err := req.GET()
 	if err != nil {
-		log.Error().Err(err).Str("cmp", "imageproxy").Str("url", req.sourceURL).Msg("unable to create GET request")
+		s.log.Error().Err(err).Str("url", req.SourceURL).Msg("unable to create GET request")
+		s.m.IncRemoteRequestError()
+
 		return err
 	}
 
 	resp, err := s.client.Do(proxyRequest)
 	if err != nil || resp.StatusCode != http.StatusOK {
-		log.Error().Err(err).Str("cmp", "imageproxy").Str("url", req.sourceURL).Msg("unable to make remote request")
+		s.log.Error().Err(err).Str("url", req.SourceURL).Msg("unable to make remote request")
+		s.m.IncRemoteRequestError()
+
 		return err
 	}
 
@@ -146,80 +269,79 @@ func (s *Service) Do(ctx context.Context, req *Request, w io.Writer) error {
 	var buf bytes.Buffer
 	tee := io.TeeReader(resp.Body, &buf)
 
-	// copy to response writer
-	_, err = io.Copy(w, tee)
-	if err != nil {
-		log.Error().Err(err).Str("cmp", "imageproxy").Msg("error copying remote image")
+	contentType := resp.Header.Get("Content-Type")
 
-		return err
-	}
-
-	if strings.HasPrefix(resp.Header.Get("Content-Type"), "text/xml") && bytes.Contains(buf.Bytes(), []byte("adlibXML")) {
+	if strings.HasPrefix(contentType, "text/xml") && bytes.Contains(buf.Bytes(), []byte("adlibXML")) {
 		// don't cache adlib error messages
-		log.Warn().Str("cmp", "imageproxy").Str("url", req.sourceURL).Msg("adlib error retrieving image")
+		s.log.Warn().Str("url", req.SourceURL).Msg("adlib error retrieving image")
+		s.m.IncRemoteRequestError()
+
 		return fmt.Errorf("unable to retrieve adlib result")
 	}
-	// check for adlib error when content type is xml
 
-	err = req.Write(cachePath, &buf)
+	// check for adlib error when content type is xml
+	err = req.Write(req.downloadedSourcePath(), tee)
 	if err != nil {
 		// do not return error here or cache write error
-		log.Error().Err(err).Str("cmp", "imageproxy").Msg("unable to write remote file to cache")
+		s.log.Error().Err(err).Msgf("unable to write remote file to cache; %s", err)
+
+		s.m.IncRemoteRequestError()
 	}
+
+	s.m.IncSource()
 
 	return nil
 }
 
-// create handler fuction to serve the proxied images
-func (s *Service) proxyImage(w http.ResponseWriter, r *http.Request) {
-	url := chi.URLParam(r, "*")
+func (s *Service) domainAllowed(targetURL string) (bool, error) {
+	if len(s.allowList) == 0 {
+		return true, nil
+	}
 
-	// add referer
-	if len(s.referrers) != 0 {
-		var allowed bool
+	var allowed bool
 
-		for _, referrer := range s.referrers {
-			if strings.Contains(r.Referer(), referrer) {
-				allowed = true
-				break
-			}
+	for _, target := range s.allowList {
+		u, err := url.Parse(targetURL)
+		if err != nil {
+			s.log.Error().Err(err).Str("target_url", targetURL).Msg("unable to parse target url")
+			return false, err
 		}
 
-		if !allowed {
-			http.Error(w, "not found", http.StatusNotFound)
-
-			return
+		if strings.HasSuffix(u.Host, target) {
+			allowed = true
+			break
 		}
 	}
 
-	req, err := NewRequest(
-		url,
-		SetRawQueryString(r.URL.RawQuery),
-	)
-	if err != nil {
-		log.Error().Err(err).Str("cmp", "imageproxy").Str("url", url).Msg("unable to create proxy request")
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-
-		return
+	if !allowed {
+		return false, nil
 	}
 
-	if len(s.blacklist) != 0 {
-		for _, uri := range s.blacklist {
-			if strings.Contains(req.sourceURL, uri) {
-				http.Error(w, "not found", http.StatusNotFound)
-				return
-			}
-		}
-	}
-
-	err = s.Do(r.Context(), req, w)
-	if err != nil {
-		log.Error().Err(err).Str("cmp", "imageproxy").Str("url", req.sourceURL).Msg("unable to make proxy request")
-		http.Error(w, err.Error(), http.StatusInternalServerError)
-
-		return
-	}
+	return true, nil
 }
+
+func (s *Service) reffererAllowed(referrer string) bool {
+	if len(s.referrers) == 0 {
+		return true
+	}
+
+	var allowed bool
+
+	for _, allowedReferrer := range s.referrers {
+		if strings.Contains(referrer, allowedReferrer) {
+			allowed = true
+			break
+		}
+	}
+
+	return allowed
+}
+
+// func (s *Service) SetOrganizationService(svc domain.Service) error {
+// // s.organizations = svc
+// // do nothing because can't set itself
+// return nil
+// }
 
 func (s *Service) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 	// needed to implement ikuzo service interface
@@ -228,3 +350,8 @@ func (s *Service) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 func (s *Service) Shutdown(ctx context.Context) error {
 	return nil
 }
+
+// func (s *Service) SetServiceBuilder(b *domain.ServiceBuilder) {
+// s.log = b.Logger.With().Str("svc", "imageproxy").Logger()
+// s.orgs = b.Orgs
+// }